style customer notice
