+++ Code/Client/users/src/components/molecules/carRental/priceCard/rentalPriceCard.module.css
@@ -1,6 +1,6 @@
 .contanier{
     width: 20rem;
-    padding-top: 6.3rem;
+    /* padding-top: 6.3rem; */
 }
 
 .card{
