+++ Code/Client/users/src/pages/CarRental/bookingPage/rentalBookingPage.module.css
@@ -1,0 +1,18 @@
+.container {
+    display: flex;
+    flex-direction: row;
+    padding: 40px 40px;
+    background-color: rgb(226, 222, 222);
+    height: 100%;
+    gap: 40px;
+}
+
+.bookingContainer {
+    display: flex;
+    flex-wrap: wrap;
+    width: 56rem;
+}
+
+.Notice {
+    height: 80rem;
+}
